requirements.txt
django==3.2.10
GitPython==3.1.18
firebase==3.0.1
firebase-admin==5.2.0
requests==2.27.1
typesense==0.13.1
djangorestframework==3.13.1
django-environ==0.8.1
drf-spectacular==0.26.1
